+++ src/poc-data/extracted_course_works.csv
@@ -23,3 +23,8 @@ TEST,Assignment 1: Art Analysis,2024-01-15,2024-01-30,15
 TEST,Midterm Exam,2024-02-20,2024-02-20,35
 TEST,Final Project: Creative Artwork,Not Found,2024-04-10,50
 
+4C03,Assignment 1: Deconstructed Essay,Not Found,2024-10-16,25
+4C03,Test 2,Not Found,2024-11-27,25
+4C03,Assignment 2: Essay,Not Found,2024-12-07,25
+test2,Assignment 1: Deconstructed Essay,Not Found,2024-10-16,25
+test2,Assignment 2: Essay,Not Found,2024-12-07,25
