+++ src/poc-data/mock_course_info.csv
@@ -14,3 +14,7 @@ Optional Reference Textbook:
 - Tuesday: 11:30 am – 12:20 pm, Location: JHE 326H""",Not Found,"""Computer networking is a rapidly advancing field. The Internet is already an integral part of society. It is therefore important for computer science, computer engineering, and software engineering students to be familiar with the fundamentals as well as practices of computer networking. This undergraduate course will emphasize the basic principles of computer networking with a focus on the architecture, algorithms, and protocols of the Internet. Topics include physical networks, TCP/IP protocols, switching methods, network layering and components, network services, information security, computer and network security threats, defense mechanisms, and encryption.""",Not Found,In case of missed evaluations please refer to the Policy on Requests for Relief for Missed Academic Term Work process document posted on the MOSAIC. Missing assignments with MSAF approvals will be counted toward the total grade of assignments of the SAME category.
 SFWRENG 4G06,4G06-syllabus.pdf,Dr. Spencer Smith,smiths@mcmaster.ca,By appointment,No required textbook.,Not Found,Not Found,"""Christopher Schankula (schankuc@mcmaster.ca); Samuel Crawford (crawfs1@mcmaster.ca); Lyuyang Wang (wangl307@mcmaster.ca)""","""Software Design IV — Capstone Design Project""","""The SE capstone provides students with an opportunity to integrate what they have learned in earlier courses, deepen their understanding of that material, extend their area of knowledge and to apply their knowledge and skills in a realistic simulation of professional experience. The emphasis of this course is on design and the design process.""",": In the event of an absence for medical or other reasons, students should review and follow the Academic Regulation in the Undergraduate Calendar “Requests for Relief for Missed Academic Term Work”."
 TEST,TEST-syllabus.pdf,Not Found,Not Found,Not Found,Not Found,Not Found,Not Found,Not Found,Not Found,Not Found,Not Found
+4C03,4C03-syllabus.pdf,Beatrice Fletcher,fletchba@mcmaster.ca,Tuesdays 2-3pm or by appointment,"Required Textbook: Introduction to Anthropology - Sex, Food and Death: Custom edition for McMaster University, Oxford University Press, 2023. 
+   Available at McMaster’s Campus Store (CAD) as a print or e-text. 
+   ISBN-13: 9780190177928, eISBN-13: 0190177926",Asynchronous Online,Not Found,Not Found,"This course introduces students to the field of Anthropology and its sub-disciplines through discussing various anthropological perspectives on sex, food, illness, and death from archaeological, biological, sociocultural, and linguistic viewpoints. Students will develop a better understanding of how anthropological approaches help in understanding the human condition past and present.",Not Found,Not Found
+test2,test2-syllabus.pdf,Beatrice Fletcher,fletchba@mcmaster.ca,Tuesday 2-3pm or by appointment,"Introduction to Anthropology - Sex, Food, and Death (Custom edition for McMaster University, Oxford University Press, 2023) ISBN-13: 9780190177928, eISBN-13: 0190177926 (Note: Textbook updated as of September 2023)",Asynchronous Online,Not Found,Not Found,"This course introduces students to the field of Anthropology and its sub-disciplines through exploring topics such as sex, gender, food, illness, and death from various anthropological perspectives. Students will develop critical thinking and analytical skills to understand the human condition past and present.",Not Found,Not Found
