+++ src/templates/index.html
@@ -142,7 +142,7 @@
                 <div class="courses" id = "course-block"
                     style="display: flex; flex-wrap: wrap; justify-content: flex-start; width: 100%; border-spacing: 10px;">
                     {% for course in courses %}
-                    <a href="/course_detail_page/{{ course }}" class="course-detail-link" style="flex-basis: 256px;">
+                    <a href="/courses/course_detail_page/{{ course }}" class="course-detail-link" style="flex-basis: 256px;">
                         <div id="{{ course.0.1 }}" class=" course-detail">
                             <div class="course-image">
                                 <img style="vertical-align:middle; width: 256px;"
